+++ pysrc/whenever/_core.py
@@ -55,3 +55,11 @@ except ModuleNotFoundError as e:
     )
 
     _EXTENSION_LOADED = False
+
+MONDAY = Weekday.MONDAY
+TUESDAY = Weekday.TUESDAY
+WEDNESDAY = Weekday.WEDNESDAY
+THURSDAY = Weekday.THURSDAY
+FRIDAY = Weekday.FRIDAY
+SATURDAY = Weekday.SATURDAY
+SUNDAY = Weekday.SUNDAY